restructuring modals into a directory
