+++ app/src/app/components/modals/merge-modal/confirm-operation/confirm-operation.component.html
