+++ app/src/app/components/modals/merge-modal/defaults-file-upload/defaults-file-upload.component.scss
@@ -1,7 +1,7 @@
 // Copyright 2020 Google LLC
 
 // Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this spec except in compliance with the License.
+// you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 
 //     https://www.apache.org/licenses/LICENSE-2.0
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import '../../../../shared/constants.scss';
+@import '../../../../../shared/constants.scss';
 @import '../merge-modal.component.scss';
 
-.spec-files {
+.defaults-file {
   &-upload-input {
     display: none;
   }
@@ -23,7 +23,10 @@
   &-content {
     margin-bottom: 2rem;
     margin-top: 2rem;
-    max-width: calc(100% - 32px);
+  }
+
+  &-chip-icon {
+    margin-right: 8px;
   }
   
   &-icon {
@@ -38,17 +41,5 @@
     width: 136px;
     background-color: $material-grey-400;
     border-radius: 100%;
-  }
-
-  &-input {
-    &-container {
-      display: flex;
-      flex-direction: column;
-      align-items: center;
-
-      #spec-files-upload-error {
-        margin-top: 1rem;
-      }
-    }
   }
 }
