+++ app/src/app/components/modals/merge-modal/merge-modal.component.scss
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import '../../../shared/constants.scss';
+@import '../../../../shared/constants.scss';
 
 .modal-container {
   position: relative;
