+++ app/src/app/components/modals/merge-modal/merge-modal.component.spec.ts
