+++ app/src/app/components/modals/merge-modal/resolve-conflict/resolve-conflict.component.html
