+++ app/src/app/components/modals/merge-modal/spec-files-upload/spec-files-upload.component.spec.ts
@@ -19,7 +19,6 @@ import { BrowserModule } from '@angular/platform-browser';
 import { MatIconModule } from '@angular/material/icon';
 import { MatChipsModule } from '@angular/material/chips';
 import { MatButtonModule } from '@angular/material/button';
-import { queryElement } from '../../../../shared/functions';
 
 describe('SpecFilesUploadComponent', () => {
   let fixture: ComponentFixture<SpecFilesUploadComponent>;
