+++ app/src/app/components/modals/merge-modal/spec-name-input/spec-name-input.component.spec.ts
