+++ app/src/app/components/modals/overlay-modal/overlay-modal.component.scss
@@ -12,17 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-@import '../../../../shared/constants.scss';
-@import '../merge-modal.component.scss';
 
-.spec-name {
-  &-form {
-    min-width: 150px;
-    max-width: 500px;
-    width: 100%;
-  }
-
-  &-field {
-    width: 100%;
-  }
-}
